backend/invoice-processor/requirements.txt
google-cloud-functions
google-cloud-storage
google-cloud-vision
google-cloud-documentai
google-cloud-aiplatform # Dùng cho Gemini API
Flask
gunicorn 